Add GetMoviesWithPostersUseCase interface
